corrects a bug on 64Bit platforms
git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1416 235a714a-1e17-0410-b3bb-947d69a96898

+++ include/mw/nl/CANNL.h
@@ -86,7 +86,7 @@ namespace famouso {
                     ~CANNL() {}
 
                     void init() {
-                        init(NodeID(0ull));
+                        init(NodeID(static_cast<unsigned long long>(0ull)));
                     }
 
                     void init(const NodeID &i) {
